mutex unlock bug

+++ trunk/src/devel/mutex.h
@@ -3,5 +3,5 @@
 
 void *mutex_create();
 void mutex_free(void *dir);
-uint8_t mutex_try_lock(uint32_t *dir);
-uint8_t mutex_try_unlock(uint32_t *dir);
+uint8_t mutex_try_lock(uint8_t *dir);
+uint8_t mutex_try_unlock(uint8_t *dir);
